fxns_create_dataset/requirements.txt
azure-core==1.30.2
azure-storage-file-share==12.17.0
beautifulsoup4==4.12.3
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.9.0
colorama==0.4.6
cryptography==43.0.0
deprecation==2.1.0
filelock==3.15.4
firebase-admin==6.5.0
firebase-functions==0.4.2
Flask==3.0.3
Flask-Cors==4.0.1
functions-framework==3.8.1
gdown==5.2.0
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-events==0.12.0
google-resumable-media==2.7.2
googleapis-common-protos==1.64.0
grpcio==1.66.0
grpcio-status==1.66.0
h5py==3.11.0
httplib2==0.22.0
idna==3.8
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgpack==1.0.8
npTDMS==1.9.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
proto-plus==1.24.0
protobuf==5.27.3
psp_liquids_daq_parser==0.0.20
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.16.0
soupsieve==2.6
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
watchdog==5.0.0
Werkzeug==3.0.4
azure-cli==2.61.0